+++ CHANGELOG.md
@@ -16,3 +16,4 @@
 - CMD+SHIFT+] on Mac / CMD+PGDN on Windows/Linux (next tab)
 - CMD+SHIFT+[ on Mac / CMD+PGUP on Windows/Linux (previous tab)
 - Use Chromium Engine
+- Ability to change engine on any platform through preferences